StyleGuru-Backend/requirements.txt
fastapi==0.115.0
uvicorn>=0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
cloudinary==1.41.0
anthropic==0.39.0
httpx==0.27.2
pillow>=11.2.0
torch>=2.2.0
torchvision>=0.17.0
